+++ Products/CMFPlone/interfaces/__init__.py
@@ -9,6 +9,7 @@ from controlpanel import IControlPanel
 from controlpanel import IDateAndTimeSchema
 from controlpanel import IEditingSchema
 from controlpanel import ILanguageSchema
+from controlpanel import IFilterSchema
 from controlpanel import IMaintenanceSchema
 from controlpanel import IMailSchema
 from controlpanel import IMarkupSchema